python/requirements.txt
aws_cdk.aws_servicecatalog==1.119.0
aws_cdk.core==1.119.0
